requirements.txt
decorator==4.4.0
entrypoints==0.3
json5==0.8.5
jsonschema==3.0.2
numpy==1.17.0
opencv-contrib-python==4.1.0.25
pandas==0.25.0
Pillow==5.4.1
pipenv==11.9.0
pytesseract==0.2.7
python-dateutil==2.7.3
python-telegram-bot==11.1.0
requests==2.21.0
requests-unixsocket==0.1.5
Unidecode==1.1.1
urllib3==1.24.1
virtualenv==15.1.0
virtualenv-clone==0.3.0

